Bug-fixes

+++ frontend/src/components/Edit.js
@@ -7,17 +7,22 @@ import workoutService from '../services/workout'
 import Container from 'react-bootstrap/Container'
 import Row from 'react-bootstrap/Row'
 import Col from 'react-bootstrap/Col'
+import Form from 'react-bootstrap/Form'
 
-const Edit = ({ exercise, sets, repetitions, weight, id }) => {
+const Edit = ({ exercise, sets, repetitions, weight, id, setUpdated, updated }) => {
   const [show, setShow] = useState(false);
   const [editSets, setEditSets] = useState(sets)
   const [editRepetitions, setEditRepetitions] = useState(repetitions)
   const [editWeight, setEditWeight] = useState(weight)
+  const [editName, setEditName] = useState(exercise)
+  const [acceptedName, setAcceptedName] = useState(exercise)
+  const [openEditName, setOpenEditName] = useState(false)
 
   const handleShow = () => setShow(true);
   const handleSets = (e) => setEditSets(e.target.value)
   const handleRepetitions = (e) => setEditRepetitions(e.target.value)
   const handleWeight = (e) => setEditWeight(e.target.value)
+  const handleName = (e) => setEditName(e.target.value)
 
   const handleClose = () => {
     setShow(false)
@@ -26,20 +31,38 @@ const Edit = ({ exercise, sets, repetitions, weight, id }) => {
     setEditWeight(weight)
   } 
 
-  const handleSave = () => {
+  const handleSave = async () => {
 
     const newData = {
       sets: Number(editSets),
       repetitions: Number(editRepetitions),
       weight: Number(editWeight)
     }
-    editService.editValues(id, newData)
+    await editService.editValues(id, newData)
     setShow(false)
+    setUpdated(!updated)
   }
 
-  const handleDelete = () => {
-    workoutService.deleteWorkout(id)
-    setShow(false)
+  const handleDelete = async () => {
+    if(window.confirm(`Are you sure you want to delete exercise '${exercise}'?`)) {
+      await workoutService.deleteWorkout(id)
+      setShow(false)
+      setUpdated(!updated)
+    }
+  }
+
+  const handleEditButton = () => {
+    setOpenEditName(!openEditName)
+    console.log(openEditName)
+  }
+
+  const handleCancel = () => {
+    setOpenEditName(!openEditName)
+  }
+
+  const handleOk = () => {
+    setAcceptedName(editName)
+    setOpenEditName(!openEditName)
   }
 
   return (
@@ -49,8 +72,20 @@ const Edit = ({ exercise, sets, repetitions, weight, id }) => {
       <Modal show={show} onHide={handleClose}>
         <Modal.Header closeButton>
           <Modal.Title>
-            {exercise} 
-            <Button style={{color:"red"}} variant="btn btn-link btn-sm" onClick={handleDelete}>delete</Button>
+
+          {openEditName && 
+          <div>
+            <input className="form-control"  placeholder="Name" value={editName} onChange={handleName}/>
+            <Button style={{color:"black"}} variant="btn btn-link btn-sm" onClick={handleOk}>ok</Button>
+            <Button style={{color:"black"}} variant="btn btn-link btn-sm" onClick={handleCancel}>cancel</Button>
+          </div>
+          }
+
+          {!openEditName && <>{acceptedName} <svg style={{paddingLeft: 2, color:"black"}} onClick={handleEditButton} xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-pencil-square" viewBox="0 0 16 16">
+            <path d="M15.502 1.94a.5.5 0 0 1 0 .706L14.459 3.69l-2-2L13.502.646a.5.5 0 0 1 .707 0l1.293 1.293zm-1.75 2.456l-2-2L4.939 9.21a.5.5 0 0 0-.121.196l-.805 2.414a.25.25 0 0 0 .316.316l2.414-.805a.5.5 0 0 0 .196-.12l6.813-6.814z"/>
+            <path fillRule="evenodd" d="M1 13.5A1.5 1.5 0 0 0 2.5 15h11a1.5 1.5 0 0 0 1.5-1.5v-6a.5.5 0 0 0-1 0v6a.5.5 0 0 1-.5.5h-11a.5.5 0 0 1-.5-.5v-11a.5.5 0 0 1 .5-.5H9a.5.5 0 0 0 0-1H2.5A1.5 1.5 0 0 0 1 2.5v11z"/>
+          </svg></>}
+          {!openEditName && <Button style={{color:"red"}} variant="btn btn-link btn-sm" onClick={handleDelete}>delete</Button>}
           </Modal.Title>
         </Modal.Header>
         <Modal.Body className="show-grid">
